feat(chat): implement file sharing and downloading functionality for files in chats
